PORTALS-2421: remove original AccountSetting component
