+++ src/App.tsx
@@ -17,11 +17,10 @@ import { TermsOfUsePage } from 'components/TermsOfUsePage'
 import TopNavBar from 'components/TopNavBar'
 import { ProfileValidation } from 'components/ProfileValidation'
 import { signOut } from 'synapse-react-client/dist/utils/SynapseClient'
-// import AccountSettings from 'components/AccountSettings'
 import { ResetPassword } from 'components/ResetPassword'
 import { ProfilePage } from 'components/ProfilePage'
 import { CertificationQuiz } from 'components/CertificationQuiz'
-import { AccountSettings2 } from 'components/AccountSettingsV2'
+import { AccountSettings } from 'components/AccountSettings'
 
 const App: React.FC = () => {
   return (
@@ -59,8 +58,7 @@ const App: React.FC = () => {
                         } else if (path === '/authenticated/signTermsOfUse') {
                           return <TermsOfUsePage />
                         } else if (path === '/authenticated/myaccount') {
-                          // return <AccountSettings/>
-                          return <AccountSettings2/>
+                          return <AccountSettings/>
                         } else if(path ==='/authenticated/myprofile') {
                           return <ProfilePage/>
                         } else if (path === '/authenticated/certificationquiz') {
