+++ src/style/_AccountSetting.scss
@@ -1,70 +1,3 @@
-#body .AccountSettings{
-    min-height: 100vh;
-    background: rgb(237, 241, 245);
-    padding-top: 40px;
-    font-size: 18px;
-
-    .container{
-        min-height: 72vh;
-        padding: 64px;
-        background: #FEFEFE;
-        box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.25);
-        display: flex;
-        flex-direction: column;
-        .verified-img-container{
-            text-align: center;
-            .verified-text{
-                margin-top: 16px;
-            }
-        }
-
-        .grid-container{
-            display: grid;
-            grid-template-columns: 1fr 3fr;
-            row-gap: 1ch;
-            .edit-cell{
-                grid-column: 1/-1;
-            }
-            .label-cell{
-                grid-column: 1;
-                font-weight: 700;
-            }
-        }
-    }
-
-    .ValidationStateContainer{
-        border: solid;
-        border-color: #4DB3B9;
-        padding: 24px;
-        width: 80%;
-        margin: auto;
-    }
-
-    .center-button{
-        text-align: center;
-        margin: 4px 0;
-    }
-
-    .account-setting-footer{
-        color: #000;
-        align-self: center;
-        margin-top: auto;
-    }
-}
-
-#body .WelcomeScreenModal {
-    .startAccountIconContainer {
-        text-align: center;
-        margin: 40px;
-    }
-    .ButtonContainer {
-        padding: 20px 0px;
-    }
-}
-.BadgeIcon {
-    margin-right: 5px;
-}
-
 .account-setting-panel{
     background-color: #fff;
     margin: 30px 0;
